# Added -zstd_implementation=cgo option.