--- a/cache/httpproxy/BUILD.bazel
+++ b/cache/httpproxy/BUILD.bazel
@@ -21,6 +21,7 @@ go_test(
         "//cache:go_default_library",
         "//cache/disk:go_default_library",
         "//cache/disk/casblob:go_default_library",
+        "//cache/disk/zstdimpl:go_default_library",
         "//utils:go_default_library",
     ],
 )