--- a/config/config.go
+++ b/config/config.go
@@ -52,6 +52,7 @@ type Config struct {
 	Dir                         string                    `yaml:"dir"`
 	MaxSize                     int                       `yaml:"max_size"`
 	StorageMode                 string                    `yaml:"storage_mode"`
+	ZstdImplementation          string                    `yaml:"zstd_implementation"`
 	HtpasswdFile                string                    `yaml:"htpasswd_file"`
 	TLSCaFile                   string                    `yaml:"tls_ca_file"`
 	TLSCertFile                 string                    `yaml:"tls_cert_file"`
@@ -85,7 +86,7 @@ var defaultDurationBuckets = []float64{.5, 1, 2.5, 5, 10, 20, 40, 80, 160, 320}
 
 // newFromArgs returns a validated Config with the specified values, and
 // an error if there were any problems with the validation.
-func newFromArgs(dir string, maxSize int, storageMode string,
+func newFromArgs(dir string, maxSize int, storageMode string, zstdImplementation string,
 	host string, port int, grpcPort int,
 	profileHost string, profilePort int,
 	htpasswdFile string,
@@ -118,6 +119,7 @@ func newFromArgs(dir string, maxSize int, storageMode string,
 		Dir:                         dir,
 		MaxSize:                     maxSize,
 		StorageMode:                 storageMode,
+		ZstdImplementation:          zstdImplementation,
 		HtpasswdFile:                htpasswdFile,
 		MaxQueuedUploads:            maxQueuedUploads,
 		NumUploaders:                numUploaders,
@@ -169,6 +171,7 @@ func newFromYamlFile(path string) (*Config, error) {
 func newFromYaml(data []byte) (*Config, error) {
 	c := Config{
 		StorageMode:            "zstd",
+		ZstdImplementation:     "go",
 		NumUploaders:           100,
 		MaxQueuedUploads:       1000000,
 		MaxBlobSize:            math.MaxInt64,
@@ -204,6 +207,9 @@ func validateConfig(c *Config) error {
 
 	if c.StorageMode != "zstd" && c.StorageMode != "uncompressed" {
 		return errors.New("storage_mode must be set to either \"zstd\" or \"uncompressed\"")
+	}
+	if c.ZstdImplementation != "go" && c.ZstdImplementation != "cgo" {
+		return errors.New("zstd_implementation must be set to either \"go\" or \"cgo\", got: " + c.ZstdImplementation)
 	}
 
 	proxyCount := 0
@@ -366,6 +372,7 @@ func get(ctx *cli.Context) (*Config, error) {
 		ctx.String("dir"),
 		ctx.Int("max_size"),
 		ctx.String("storage_mode"),
+		ctx.String("zstd_implementation"),
 		ctx.String("host"),
 		ctx.Int("port"),
 		ctx.Int("grpc_port"),