Allows to select cgo implementation of zstd if compiled in.

The cgo implementation is more performant than pure go implementation,
see https://github.com/buchgr/bazel-remote/issues/482 for actual benchmarks and dicussion.

Stats show that with cgo implementation the overhead of zstd compression itself is much lower -  40% vs 160% for downloading. It increases possible bandwidth for downloading from bazel-remote from 100MB/s to 190MB/s per single cpu the bazel-remote runs on, so roughly 2x.

For uploading the difference is smaller - 310% vs 490%, the overhead is big mainly because a) compression is slower than decompression in zstd case, and b) bazel-cache needs to perform also sha256 computation, which from basic observations is similarly expensive as zstd compression at faster levels. It increases upload of whole bazel-remote from 45MB/s to 64MB/s.


 See https://github.com/glukasiknuro/go-zstd-benchmarks/tree/master/bazel-remote-load-test#results---download